% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cE.}
\alias{cE.}
\alias{ccov.}
\alias{cvar.}
\title{Conditional expectation}
\usage{
cE.(x, g, p)

cvar.(x, g, p)

ccov.(x, y, g, p)
}
\arguments{
\item{x}{A numeric vector giving the population}

\item{g}{An atomic vector on which to condition}

\item{p}{weights}

\item{y}{another vector}
}
\value{
The conditional expectation of \code{x}.
}
\description{
Conditional expectation
}
\examples{
n <- 100
x <- rnorm(n)
y <- rnorm(n)
g <- rep(letters[1:10], 10)

# law of total expectation
E.(x)
E.(cE.(x, g))

# law of total variance
var.(x)
E.(cvar.(x, g)) + var.(cE.(x, g))

# law of total covariance
cov.(x, y)
E.(ccov.(x, y, g)) + cov.(cE.(x, g), cE.(y, g))
}

