% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cm.}
\alias{cm.}
\title{General population-level mixed central moment}
\usage{
cm.(df, p)
}
\arguments{
\item{df}{A data frame of numeric vectors of the same length.}

\item{p}{An optional vector of weights.  If missing, equal
weights are used.}
}
\value{
The mixed central moment of the variables in \code{df}.
}
\description{
Mixed central moment of an arbitrary number of vectors in a multivariate population.
}

