\name{cor.}
\alias{cor.}
\title{Population-level correlation}
\usage{
  cor.(x, y, p)
}
\arguments{
  \item{x}{A numeric vector giving the first variable of
  the population.}

  \item{y}{A numeric vector giving the second variable of
  the population.}

  \item{p}{An optional vector of weights.  If missing,
  equal weights are used.}
}
\value{
  The correlation of \code{x} \code{y}.
}
\description{
  Correlation of two vectors giving the elements of a
  bivariate population.
}

