% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coskew.}
\alias{coskew.}
\title{Population-level coskewness}
\usage{
coskew.(x, y, z, p)
}
\arguments{
\item{x}{A numeric vector giving the first variable of the population.}

\item{y}{A numeric vector giving the second variable of the population.}

\item{z}{A numeric vector giving the third variable of the population.}

\item{p}{An optional vector of weights.  If missing, equal
weights are used.}
}
\value{
The coskewness of \code{x} and \code{y} and \code{z}.  This
	is \code{E.((x - E.(x))*(y - E.(y))*(z - E.(z)))}
}
\description{
Coskewness of three vectors giving the elements of a trivariate
population.  Note that this is the unstandardized third mixed central
moment.
}

