% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dev2.}
\alias{dev2.}
\title{Population-level squared deviation}
\usage{
dev2.(x, p)
}
\arguments{
\item{x}{A numeric vector giving the population.}

\item{p}{An optional vector of weights.  If missing, equal
weights are used.}
}
\value{
The deviations of \code{x}.
}
\description{
Squared deviations of the elements of a vector from their population mean.
}

