\name{E.}
\alias{E.}
\title{Population-level expectation}
\usage{
  E.(x, p)
}
\arguments{
  \item{x}{A numeric vector giving the population.}

  \item{p}{An optional vector of weights.  If missing,
  equal weights are used.}
}
\value{
  The expectation of \code{x}.
}
\description{
  Expectation of a vector giving the elements of a
  population.
}

