% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Edev2.}
\alias{Edev2.}
\title{Population-level mean squared deviation}
\usage{
Edev2.(x, p)
}
\arguments{
\item{x}{A numeric vector giving the population.}

\item{p}{An optional vector of weights.  If missing, equal
weights are used.}
}
\value{
The deviations of \code{x}.
}
\description{
Mean squared deviations of the elements of a vector from their population mean.
}

