% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Edist}
\alias{Edist}
\title{Expected value of a distribution}
\usage{
Edist(dfunc, lower = -Inf, upper = Inf, support, ...)
}
\arguments{
\item{dfunc}{Function that takes a random variable as its first
argument and returns its probability (if \code{support} in not
missing or probability density (if \code{support} is missing).}

\item{lower}{Lower limit of the continuous \code{support} of the
distribution (only has an effect if \code{support} is missing).}

\item{upper}{Upper limit of the continuous \code{support} of the
distribution (only has an effect if \code{support} is missing).}

\item{support}{Values at which to evaluate a distribution with
discrete support.}

\item{...}{Additional arguments to pass to \code{dfunc}.}
}
\value{
The value of the expected value of \code{dfunc} (if
 \code{support} is not missing) or an approximation to the
 expected value supplied by \code{\link{integrate}} (if
 \code{support} is missing).
}
\description{
Calculate or approximate the expected value of a distribution.
If \code{support} is missing, it is assumed that the support is
continuous and \code{\link{integrate}} is used to approximate
the expected value.  If the support is countably infinite, then
\code{support} should be chosen to reduce the error associated
numerically approximating infinite sums.
}
\examples{
Edist(dnorm) # exactly 0
Edist(dpois, lambda = 1, support = 0:3) # bad approximation to 1
Edist(dpois, lambda = 1, support = 0:100) # (numerically) exactly 1
Edist(dbinom, size = 4, prob = 0.51, support = 0:4)
Edist(dgamma, lower = 0, shape = 2.3, scale = 0.2)
Edist(sin, lower = -1, upper = 1)
}

