\docType{package}
\name{popmoments}
\alias{package-popmoments}
\alias{popmoments}
\alias{popmoments-package}
\title{Population-level moments}
\description{
  Simple functions for calculating the population-level
  moments.
}
\details{
  The expectation (\code{\link{E.}}), variance
  (\code{\link{var.}}), covariance (\code{\link{cov.}}),
  correlation (\code{\link{cor.}}), and regression
  (\code{\link{reg.}}) operators are provided. Note that
  each function ends in a \code{.} to distinguish them from
  other standard functions with similar names.
}

