% Generated by roxygen2 (4.0.1): do not edit by hand
\name{R2.}
\alias{R2.}
\title{Population-level R-squared}
\usage{
R2.(x, y, p)
}
\arguments{
\item{x}{A numeric vector giving the population.}

\item{y}{A numeric vector giving the population.}

\item{p}{An optional vector of weights.  If missing, equal
weights are used.}
}
\value{
The R-squared of \code{x} and \code{y}.
}
\description{
R-squared
}

