% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sd.}
\alias{sd.}
\title{Population-level standard deviation}
\usage{
sd.(x, p)
}
\arguments{
\item{x}{A numeric vector giving the population.}

\item{p}{An optional vector of weights.  If missing, equal
weights are used.}
}
\value{
The standard deviation of \code{x}.
}
\description{
Standard deviation of a vector giving the elements of a population.
}

