% Generated by roxygen2 (4.0.1): do not edit by hand
\name{skew.}
\alias{skew.}
\title{Population-level skewness}
\usage{
skew.(x, p)
}
\arguments{
\item{x}{A numeric vector giving the first variable of the population.}

\item{p}{An optional vector of weights.  If missing, equal
weights are used.}
}
\value{
The skewness of \code{x}.  This
 is \code{E.((x - E.(x))^3)}
}
\description{
Skewness of a vector giving the elements of a population. Note that this
is the unstandardized third central moment.
}

