% Generated by roxygen2 (4.0.1): do not edit by hand
\name{z.}
\alias{z.}
\title{Population-level z-score}
\usage{
z.(x, p)
}
\arguments{
\item{x}{A numeric vector giving the population.}

\item{p}{An optional vector of weights.  If missing, equal
weights are used.}
}
\value{
The z-scores of \code{x}.
}
\description{
Transform a vector into z-scores using population mean and standard deviation.
}

